Fixed an issue with sortedArrayToBST function
The function would properly balance a sorted array. When I first wrote
the function, the BST looked/was balanced but when I would add more
nodes and then call the rebalance method the BST would NOT place some
nodes or have them misplaced. (ie. 9 would be the root node but 8 would
be to the right of 9 even though it was less than 9).